Render Markdown in project descriptions. Most Markdown syntax is supported with the exception of tables, reference links, abbreviations. Images can be embedded but will not be auto-resized to fit width yet. There's support for previewing in the edit project screen, though the UI is not great (floating always on top window).
Re-write the "Read More" support to work with the Markdown support and have a whizzier animation.

Introduce Kotlin to the project. In future I will convert more code to Kotlin, but for now, just the Markdown processing code is using it.

Currently this is unshippable because it turns out that Pegdown isn't compatible with Java 8. All it needs is a dependency bump but this would result in Scala version skew: I'm calling for help from the old maintainer.

Resolves issue #108