Big refactor / rewrite / convert to Kotlin of backend code, part three
Rewrite the startup sequence to be simpler, clearer and faster. Move away from WalletAppKit, which needs a rewrite anyway. Replace it with BitcoinBackend. We still use WAK on the server where startup time is less important.

There's still lots of potential for startup time optimisation and now it's a lot easier to do.